Refactor (mostly split) type inference code
