+++ nemo/src/program_analysis/analysis.rs
@@ -1,27 +1,21 @@
 use std::collections::{HashMap, HashSet};
 
-use nemo_physical::{aggregates::operation::AggregateOperation, management::database::ColumnOrder};
+use nemo_physical::management::database::ColumnOrder;
 
 use crate::{
     error::Error,
     model::chase_model::{ChaseProgram, ChaseRule},
     model::{
-        chase_model::{ChaseAtom, PrimitiveAtom, VariableAtom, AGGREGATE_VARIABLE_PREFIX},
-        types::error::TypeError,
-        Constraint, DataSource, Identifier, PrimitiveTerm, PrimitiveType, Term, TypeConstraint,
-        Variable,
+        chase_model::{ChaseAtom, PrimitiveAtom, VariableAtom},
+        Constraint, DataSource, Identifier, PrimitiveTerm, PrimitiveType, Term, Variable,
     },
-    util::labeled_graph::LabeledGraph,
 };
 
-use super::variable_order::{
-    build_preferable_variable_orders, BuilderResultVariants, VariableOrder,
+use super::{
+    type_inference::infer_types,
+    variable_order::{build_preferable_variable_orders, BuilderResultVariants, VariableOrder},
 };
 
-use petgraph::{
-    visit::{Dfs, EdgeFiltered},
-    Directed,
-};
 use thiserror::Error;
 
 /// Contains useful information for a (existential) rule
@@ -114,7 +108,7 @@ fn get_predicates<Atom: ChaseAtom>(atoms: &[Atom]) -> HashSet<Identifier> {
     atoms.iter().map(|a| a.predicate()).collect()
 }
 
-fn get_fresh_rule_predicate(rule_index: usize) -> Identifier {
+pub(super) fn get_fresh_rule_predicate(rule_index: usize) -> Identifier {
     Identifier(format!(
         "FRESH_HEAD_MATCHES_IDENTIFIER_FOR_RULE_{rule_index}"
     ))
@@ -209,39 +203,18 @@ fn analyze_rule(
     promising_variable_orders: Vec<VariableOrder>,
     promising_column_orders: &[HashMap<Identifier, HashSet<ColumnOrder>>],
     rule_index: usize,
+    variable_types: HashMap<Variable, PrimitiveType>,
     type_declarations: &HashMap<Identifier, Vec<PrimitiveType>>,
 ) -> RuleAnalysis {
     let num_existential = count_distinct_existential_variables(rule);
 
-    let mut variable_types: HashMap<Variable, PrimitiveType> = HashMap::new();
-    let mut rule_all_predicates = Vec::<Identifier>::new();
-
-    for atom in rule.head() {
-        rule_all_predicates.push(atom.predicate());
-
-        for (term_position, term) in atom.terms().iter().enumerate() {
-            if let PrimitiveTerm::Variable(variable) = term {
-                variable_types.entry(variable.clone()).or_insert(
-                    type_declarations
-                        .get(&atom.predicate())
-                        .expect("Every predicate should have received type information.")
-                        [term_position],
-                );
-            }
-        }
-    }
-    for atom in rule.all_body() {
-        rule_all_predicates.push(atom.predicate());
-
-        for (term_position, variable) in atom.terms().iter().enumerate() {
-            variable_types.entry(variable.clone()).or_insert(
-                type_declarations
-                    .get(&atom.predicate())
-                    .expect("Every predicate should have received type information.")
-                    [term_position],
-            );
-        }
-    }
+    let rule_all_predicates: Vec<_> = rule
+        .all_body()
+        .cloned()
+        .map(Into::into)
+        .chain(rule.head().iter().cloned())
+        .map(|a| a.predicate())
+        .collect();
 
     let (existential_aux_rule, existential_aux_order, existential_aux_types) =
         if num_existential > 0 {
@@ -281,127 +254,6 @@ fn analyze_rule(
     }
 }
 
-/// Identifies a position within a predicate.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct PredicatePosition {
-    predicate: Identifier,
-    position: usize,
-}
-
-impl PredicatePosition {
-    /// Create new [`PredicatePosition`].
-    pub fn new(predicate: Identifier, position: usize) -> Self {
-        Self {
-            predicate,
-            position,
-        }
-    }
-}
-
-/// Edge Types in Position Graph for Type Analysis
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum PositionGraphEdge {
-    WithinBody,
-    BodyToHeadSameVariable,
-    /// Describes data flow from an aggregate input variable to and aggregate output variable, where the output variable has a static type (see [`AggregateOperation::static_output_type`]).
-    /// This has no impact on type propagation from input to output, but is there for completeness sake, as it still describes some data flow.
-    BodyToHeadAggregateStaticOutputType,
-    /// Describes data flow from an aggregate input variable to and aggregate output variable, where the aggregate output variable has the same type as the input variable (see [`AggregateOperation::static_output_type`]).
-    BodyToHeadAggregateNonStaticOutputType,
-}
-/// Graph that represents a prioritization between rules.
-pub type PositionGraph = LabeledGraph<PredicatePosition, PositionGraphEdge, Directed>;
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-enum TypeRequirement {
-    Hard(PrimitiveType),
-    Soft(PrimitiveType),
-    None,
-}
-
-impl TypeRequirement {
-    fn stricter_requirement(self, other: Self) -> Option<Self> {
-        match self {
-            Self::Hard(t1) => match other {
-                Self::Hard(t2) => (t1 == t2).then_some(self),
-                Self::Soft(t2) => (t1 >= t2).then_some(self),
-                Self::None => Some(self),
-            },
-            Self::Soft(t1) => match other {
-                Self::Hard(t2) => (t1 <= t2).then_some(other),
-                Self::Soft(t2) => t1.partial_cmp(&t2).map(|ord| match ord {
-                    std::cmp::Ordering::Equal => self,
-                    std::cmp::Ordering::Greater => self,
-                    std::cmp::Ordering::Less => other,
-                }),
-                Self::None => Some(self),
-            },
-            Self::None => Some(other),
-        }
-    }
-
-    fn replacement_type_if_compatible(self, other: Self) -> Option<Self> {
-        match self {
-            Self::Hard(t1) => match other {
-                Self::Hard(t2) => (t1 == t2).then_some(self),
-                Self::Soft(t2) => (t1 >= t2).then_some(self),
-                Self::None => Some(self),
-            },
-            Self::Soft(t1) => match other {
-                Self::Hard(t2) => t1.partial_cmp(&t2).map(|ord| match ord {
-                    std::cmp::Ordering::Equal => self,
-                    std::cmp::Ordering::Greater => self,
-                    std::cmp::Ordering::Less => Self::Soft(t2),
-                }),
-                Self::Soft(t2) => t1.partial_cmp(&t2).map(|ord| match ord {
-                    std::cmp::Ordering::Equal => self,
-                    std::cmp::Ordering::Greater => self,
-                    std::cmp::Ordering::Less => other,
-                }),
-                Self::None => Some(self),
-            },
-            Self::None => match other {
-                Self::Hard(t2) => Some(Self::Soft(t2)),
-                Self::Soft(t2) => Some(Self::Soft(t2)),
-                Self::None => Some(Self::None),
-            },
-        }
-    }
-
-    fn allowed_to_merge_with(self, other: Self) -> bool {
-        match Option::<PrimitiveType>::from(self) {
-            Some(t1) => match Option::from(other) {
-                Some(t2) => t1.partial_cmp(&t2).is_some(),
-                None => true,
-            },
-            None => true,
-        }
-    }
-}
-
-impl From<TypeConstraint> for TypeRequirement {
-    fn from(value: TypeConstraint) -> Self {
-        match value {
-            TypeConstraint::None => TypeRequirement::None,
-            TypeConstraint::Exact(p) => TypeRequirement::Hard(p),
-            TypeConstraint::AtLeast(p) => TypeRequirement::Soft(p),
-            TypeConstraint::Tuple(_) => {
-                unimplemented!("currently nested type checking is not supported")
-            }
-        }
-    }
-}
-
-impl From<TypeRequirement> for Option<PrimitiveType> {
-    fn from(source: TypeRequirement) -> Self {
-        match source {
-            TypeRequirement::Hard(t) => Some(t),
-            TypeRequirement::Soft(t) => Some(t),
-            TypeRequirement::None => None,
-        }
-    }
-}
-
 /// Contains useful information about the
 #[derive(Debug)]
 pub struct ProgramAnalysis {
@@ -413,8 +265,6 @@ pub struct ProgramAnalysis {
     pub all_predicates: HashSet<(Identifier, usize)>,
     /// Logical Type Declarations for Predicates
     pub predicate_types: HashMap<Identifier, Vec<PrimitiveType>>,
-    /// Graph representing the information flow between predicates
-    pub position_graph: PositionGraph,
 }
 
 impl ChaseProgram {
@@ -432,7 +282,7 @@ impl ChaseProgram {
     }
 
     /// Collect all predicates occurring in the program.
-    fn get_all_predicates(&self) -> HashSet<(Identifier, usize)> {
+    pub(super) fn get_all_predicates(&self) -> HashSet<(Identifier, usize)> {
         let mut result = HashSet::<(Identifier, usize)>::new();
 
         // Predicates in source statements
@@ -473,537 +323,6 @@ impl ChaseProgram {
         }
 
         result
-    }
-
-    fn build_position_graph(&self) -> PositionGraph {
-        let mut graph = PositionGraph::default();
-
-        for rule in self.rules() {
-            let mut variables_to_head_positions =
-                HashMap::<Variable, Vec<PredicatePosition>>::new();
-
-            for atom in rule.head() {
-                for (term_position, term) in atom.terms().iter().enumerate() {
-                    if let PrimitiveTerm::Variable(variable) = term {
-                        let predicate_position =
-                            PredicatePosition::new(atom.predicate(), term_position);
-
-                        variables_to_head_positions
-                            .entry(variable.clone())
-                            .and_modify(|e| e.push(predicate_position.clone()))
-                            .or_insert(vec![predicate_position]);
-                    }
-                }
-            }
-
-            let mut aggregate_input_to_output_variables =
-                HashMap::<Variable, Vec<(Variable, PositionGraphEdge)>>::new();
-            for aggregate in rule.aggregates() {
-                for input_variable_identifier in &aggregate.variables {
-                    let edge_label = if aggregate.aggregate_operation.static_output_type().is_some()
-                    {
-                        PositionGraphEdge::BodyToHeadAggregateStaticOutputType
-                    } else {
-                        PositionGraphEdge::BodyToHeadAggregateNonStaticOutputType
-                    };
-
-                    aggregate_input_to_output_variables
-                        .entry(input_variable_identifier.clone())
-                        .or_default()
-                        .push((aggregate.output_variable.clone(), edge_label));
-                }
-            }
-
-            let mut variables_to_last_node = HashMap::<Variable, PredicatePosition>::new();
-
-            for atom in rule.all_body() {
-                for (term_position, variable) in atom.terms().iter().enumerate() {
-                    let predicate_position =
-                        PredicatePosition::new(atom.predicate(), term_position);
-
-                    // Add head position edges
-                    {
-                        // NOTE: we connect each body position to each head position of the same variable
-                        if let Some(head_positions) = variables_to_head_positions.get(variable) {
-                            for pos in head_positions {
-                                graph.add_edge(
-                                    predicate_position.clone(),
-                                    pos.clone(),
-                                    PositionGraphEdge::BodyToHeadSameVariable,
-                                );
-                            }
-                        }
-
-                        // NOTE: we connect every aggregate input variable to it's corresponding output variable in the head
-                        if let Some(output_variable_identifiers) =
-                            aggregate_input_to_output_variables.get(variable)
-                        {
-                            for (output_variable_identifier, edge_label) in
-                                output_variable_identifiers.iter()
-                            {
-                                for pos in variables_to_head_positions
-                                    .get(&output_variable_identifier.clone())
-                                    .unwrap()
-                                {
-                                    graph.add_edge(
-                                        predicate_position.clone(),
-                                        pos.clone(),
-                                        *edge_label,
-                                    );
-                                }
-                            }
-                        }
-                    }
-
-                    // NOTE: we do not fully interconnect body positions as we start DFS from
-                    // each possible position later covering all possible combinations
-                    // nonetheless
-                    variables_to_last_node
-                        .entry(variable.clone())
-                        .and_modify(|entry| {
-                            let last_position =
-                                std::mem::replace(entry, predicate_position.clone());
-                            graph.add_edge(
-                                last_position.clone(),
-                                predicate_position.clone(),
-                                PositionGraphEdge::WithinBody,
-                            );
-                            graph.add_edge(
-                                predicate_position.clone(),
-                                last_position,
-                                PositionGraphEdge::WithinBody,
-                            );
-                        })
-                        .or_insert(predicate_position);
-                }
-            }
-
-            for constructor in rule.constructors() {
-                // Note that the head variable for constructors is unique so we can get the first entry of this vector
-                for head_position in variables_to_head_positions
-                    .get(constructor.variable())
-                    .expect("The loop at the top went through all head atoms")
-                {
-                    for term in constructor.term().primitive_terms() {
-                        if let PrimitiveTerm::Variable(body_variable) = term {
-                            let body_position = variables_to_last_node
-                                .get(body_variable)
-                                .expect("The iteration above went through all body atoms")
-                                .clone();
-
-                            graph.add_edge(
-                                body_position,
-                                head_position.clone(),
-                                PositionGraphEdge::BodyToHeadSameVariable,
-                            );
-                        }
-                    }
-                }
-            }
-
-            for constraint in rule.all_constraints() {
-                let variables = constraint
-                    .left()
-                    .variables()
-                    .chain(constraint.right().variables());
-                let next_variables = constraint
-                    .left()
-                    .variables()
-                    .chain(constraint.right().variables())
-                    .skip(1);
-
-                for (current_variable, next_variable) in variables.zip(next_variables) {
-                    let position_current = variables_to_last_node
-                        .get(current_variable)
-                        .expect("Variables in filters should also appear in the rule body")
-                        .clone();
-                    let position_next = variables_to_last_node
-                        .get(next_variable)
-                        .expect("Variables in filters should also appear in the rule body")
-                        .clone();
-
-                    graph.add_edge(
-                        position_current.clone(),
-                        position_next.clone(),
-                        PositionGraphEdge::WithinBody,
-                    );
-                    graph.add_edge(
-                        position_next,
-                        position_current,
-                        PositionGraphEdge::WithinBody,
-                    );
-                }
-            }
-        }
-
-        graph
-    }
-
-    fn infer_predicate_types(
-        &self,
-        position_graph: &PositionGraph,
-        all_predicates: &HashSet<(Identifier, usize)>,
-    ) -> Result<HashMap<Identifier, Vec<PrimitiveType>>, TypeError> {
-        let mut predicate_types: HashMap<Identifier, Vec<TypeRequirement>> = {
-            let pred_decls = self
-                .parsed_predicate_declarations()
-                .iter()
-                .map(|(pred, types)| {
-                    (
-                        pred.clone(),
-                        types
-                            .iter()
-                            .copied()
-                            .map(TypeRequirement::Hard)
-                            .collect::<Vec<_>>(),
-                    )
-                })
-                .collect::<HashMap<_, _>>();
-
-            let mut fact_and_constant_term_decls: HashMap<Identifier, Vec<Option<PrimitiveType>>> =
-                HashMap::new();
-            for fact in self.facts() {
-                let new_types: Vec<Option<PrimitiveType>> = fact
-                    .terms()
-                    .iter()
-                    .map(|c| c.primitive_type())
-                    .map(Some)
-                    .collect();
-
-                fact_and_constant_term_decls
-                    .entry(fact.predicate())
-                    .and_modify(|types| {
-                        *types = types
-                            .iter()
-                            .zip(new_types.iter())
-                            .map(|(a, b)| {
-                                a.as_ref()
-                                    .map(|a| b.as_ref().map(|b| a.max_type(b)).unwrap_or(*a))
-                                    .or(*b)
-                            })
-                            .collect();
-                    })
-                    .or_insert(new_types);
-            }
-
-            for chase_rule in self.rules() {
-                let constructors = chase_rule.constructors();
-                let constraints = chase_rule.all_constraints().collect::<Vec<_>>();
-
-                for chase_atom in chase_rule
-                    .all_body()
-                    .cloned()
-                    .map(Into::into)
-                    .chain(chase_rule.head().iter().cloned())
-                {
-                    let new_types: Vec<Option<PrimitiveType>> = chase_atom
-                        .terms()
-                        .iter()
-                        .map(|t| match t {
-                            // TODO: should we respect other things here?
-                            // variables nested in terms in constraints in particular?
-                            PrimitiveTerm::Variable(v) => constructors
-                                .iter()
-                                .filter_map(|c| (c.variable() == v).then_some(c.term()))
-                                .chain(constraints.iter().filter_map(|c| {
-                                    match c.left() {
-                                        Term::Primitive(PrimitiveTerm::Variable(var)) => (*t
-                                            == PrimitiveTerm::Variable(var.clone()))
-                                        .then_some(c.right()),
-                                        _ => match c.right() {
-                                            Term::Primitive(PrimitiveTerm::Variable(var)) => (*t
-                                                == PrimitiveTerm::Variable(var.clone()))
-                                            .then_some(c.left()),
-                                            _ => None,
-                                        },
-                                    }
-                                }))
-                                .fold(None, |acc, t| {
-                                    let b = t.primitive_type();
-                                    acc.map(|a: PrimitiveType| {
-                                        b.map(|b| a.max_type(&b)).unwrap_or(a)
-                                    })
-                                    .or(b)
-                                }),
-                            _ => t.primitive_type(),
-                        })
-                        .collect();
-
-                    fact_and_constant_term_decls
-                        .entry(chase_atom.predicate())
-                        .and_modify(|types| {
-                            *types = types
-                                .iter()
-                                .zip(new_types.iter())
-                                .map(|(a, b)| {
-                                    a.as_ref()
-                                        .map(|a| b.as_ref().map(|b| a.max_type(b)).unwrap_or(*a))
-                                        .or(*b)
-                                })
-                                .collect();
-                        })
-                        .or_insert(new_types);
-                }
-            }
-
-            let source_decls: HashMap<Identifier, Vec<TypeRequirement>> = self
-                .sources()
-                .map(|source| {
-                    (
-                        source.predicate.clone(),
-                        source
-                            .input_types()
-                            .iter()
-                            .cloned()
-                            .map(TypeRequirement::from)
-                            .collect(),
-                    )
-                })
-                .collect::<HashMap<_, _>>();
-
-            // Infer type of count aggregates, which is always `PrimitiveType::Integer`
-            let aggregate_decls = self
-                .rules()
-                .iter()
-                .flat_map(|rule| rule.head().iter().map(move |atom| (rule, atom)))
-                .map(|(rule, atom)| {
-                    (
-                        atom.predicate(),
-                        atom.terms()
-                            .iter()
-                            .map(|term| {
-                                if let PrimitiveTerm::Variable(Variable::Universal(identifier)) = term {
-                                    if identifier.name().starts_with(AGGREGATE_VARIABLE_PREFIX) {
-                                        let aggregate = rule
-                                            .aggregates()
-                                            .iter()
-                                            .find(|aggregate| aggregate.output_variable.name() == *identifier.0).expect("variable with aggregate prefix is missing an associated aggregate");
-                                        if
-                                            aggregate.aggregate_operation ==
-                                            AggregateOperation::Count
-                                        {
-                                            return TypeRequirement::Hard(PrimitiveType::Integer)
-                                        }
-                                    }
-                                }
-                                TypeRequirement::None
-                            })
-                            .collect::<Vec<_>>(),
-                    )
-                })
-                .collect::<HashMap<_, _>>();
-
-            let existential_decls = self
-                .rules()
-                .iter()
-                .flat_map(|r| r.head())
-                .map(|a| {
-                    (
-                        a.predicate(),
-                        a.terms()
-                            .iter()
-                            .map(|t| {
-                                if matches!(t, PrimitiveTerm::Variable(Variable::Existential(_))) {
-                                    TypeRequirement::Hard(PrimitiveType::Any)
-                                } else {
-                                    TypeRequirement::None
-                                }
-                            })
-                            .collect::<Vec<_>>(),
-                    )
-                })
-                .collect::<HashMap<_, _>>();
-
-            let mut predicate_types = source_decls;
-            for (pred, fact_types) in fact_and_constant_term_decls {
-                // keep track of error that might occur on conflicting type changes
-                let mut types_not_in_conflict: Result<_, _> = Ok(());
-
-                let pred_clone = pred.clone();
-
-                predicate_types
-                    .entry(pred)
-                    .and_modify(|ts| {
-                        ts.iter_mut()
-                            .zip(&fact_types)
-                            .enumerate()
-                            .for_each(|(index, (tr, ft))| match tr.stricter_requirement(ft.map(TypeRequirement::Soft).unwrap_or(TypeRequirement::None)) {
-                                Some(res) => {
-                                    *tr = res;
-                                }
-                                None => {
-                                    if let TypeRequirement::Soft(t) = tr {
-                                        *tr = ft.map(|ft| TypeRequirement::Soft(t.max_type(&ft))).unwrap_or(*tr)
-                                    } else {
-                                        types_not_in_conflict = Err(TypeError::InvalidRuleConflictingTypes(
-                                            pred_clone.0.clone(),
-                                            index,
-                                            Option::<PrimitiveType>::from(*tr).expect(
-                                                "if the type requirement is none, there is a maximum",
-                                            ),
-                                            ft.expect("if the type requirement is none, there is a maximum"),
-                                        ));
-                                    }
-                                }
-                            })
-                    })
-                    .or_insert(fact_types.into_iter().map(|ft| ft.map(TypeRequirement::Soft).unwrap_or(TypeRequirement::None)).collect());
-
-                // abort if there is a type error
-                types_not_in_conflict?
-            }
-            predicate_types.extend(pred_decls);
-            for (pred, exis_types) in existential_decls.into_iter().chain(aggregate_decls) {
-                // keep track of error that might occur on conflicting type changes
-                let mut types_not_in_conflict: Result<_, _> = Ok(());
-
-                let pred_clone = pred.clone();
-
-                predicate_types
-                    .entry(pred)
-                    .and_modify(|ts| {
-                        ts.iter_mut()
-                            .zip(&exis_types)
-                            .enumerate()
-                            .for_each(|(index, (t, et))| match t.stricter_requirement(*et) {
-                                Some(res) => {
-                                    *t = res;
-                                }
-                                None => {
-                                    types_not_in_conflict = Err(TypeError::InvalidRuleConflictingTypes(
-                                        pred_clone.0.clone(),
-                                        index,
-                                        Option::<PrimitiveType>::from(*t).expect(
-                                            "if the type requirement is none, there is a maximum",
-                                        ),
-                                        Option::<PrimitiveType>::from(*et).expect(
-                                            "if the type requirement is none, there is a maximum",
-                                        ),
-                                    ));
-                                }
-                            })
-                    })
-                    .or_insert(exis_types);
-
-                // abort if there is a type error
-                types_not_in_conflict?
-            }
-            for (predicate, arity) in all_predicates {
-                predicate_types
-                    .entry(predicate.clone())
-                    .or_insert(vec![TypeRequirement::None; *arity]);
-            }
-            predicate_types
-        };
-
-        let initial_types = predicate_types.clone();
-
-        // Propagate each type from its declaration
-        for (predicate, types) in &initial_types {
-            for (position, logical_type_requirement) in types.iter().enumerate() {
-                let predicate_position = PredicatePosition::new(predicate.clone(), position);
-
-                if let Some(start_node) = position_graph.get_node(&predicate_position) {
-                    let edge_filtered_graph = EdgeFiltered::from_fn(position_graph.graph(), |e| {
-                        *e.weight() == PositionGraphEdge::BodyToHeadSameVariable
-                            || *e.weight()
-                                == PositionGraphEdge::BodyToHeadAggregateNonStaticOutputType
-                    });
-
-                    let mut dfs = Dfs::new(&edge_filtered_graph, start_node);
-
-                    while let Some(next_node) = dfs.next(&edge_filtered_graph) {
-                        let next_position = position_graph
-                            .graph()
-                            .node_weight(next_node)
-                            .expect("The DFS iterator guarantees that every node exists.");
-
-                        let current_type_requirement = &mut predicate_types
-                            .get_mut(&next_position.predicate)
-                            .expect("The initialization step inserted every known predicate")
-                            [next_position.position];
-
-                        if let Some(replacement) = current_type_requirement
-                            .replacement_type_if_compatible(*logical_type_requirement)
-                        {
-                            *current_type_requirement = replacement;
-                        } else {
-                            return Err(TypeError::InvalidRuleConflictingTypes(
-                                next_position.predicate.0.clone(),
-                                next_position.position + 1,
-                                Option::<PrimitiveType>::from(*current_type_requirement)
-                                    .expect("if the type requirement is none, there is a maximum"),
-                                Option::<PrimitiveType>::from(*logical_type_requirement)
-                                    .expect("if the type requirement is none, there is a maximum"),
-                            ));
-                        }
-                    }
-                }
-            }
-        }
-
-        // Check compatibility of body types without overwriting
-        for (predicate, types) in &predicate_types {
-            for (position, logical_type_requirement) in types.iter().enumerate() {
-                let predicate_position = PredicatePosition::new(predicate.clone(), position);
-
-                if let Some(start_node) = position_graph.get_node(&predicate_position) {
-                    let edge_filtered_graph = EdgeFiltered::from_fn(position_graph.graph(), |e| {
-                        *e.weight() == PositionGraphEdge::WithinBody
-                    });
-
-                    let mut dfs = Dfs::new(&edge_filtered_graph, start_node);
-
-                    while let Some(next_node) = dfs.next(&edge_filtered_graph) {
-                        let next_position = position_graph
-                            .graph()
-                            .node_weight(next_node)
-                            .expect("The DFS iterator guarantees that every node exists.");
-
-                        let current_type_requirement = &predicate_types
-                            .get(&next_position.predicate)
-                            .expect("The initialization step inserted every known predicate")
-                            [next_position.position];
-
-                        if !current_type_requirement
-                            .allowed_to_merge_with(*logical_type_requirement)
-                        {
-                            // TODO: maybe just throw a warning here? (comparison of incompatible
-                            // types can be done but will trivially result in inequality)
-                            return Err(TypeError::InvalidRuleConflictingTypes(
-                                next_position.predicate.0.clone(),
-                                next_position.position + 1,
-                                Option::<PrimitiveType>::from(*current_type_requirement)
-                                    .expect("if the type requirement is none, merging is allowed"),
-                                Option::<PrimitiveType>::from(*logical_type_requirement)
-                                    .expect("if the type requirement is none, merging is allowed"),
-                            ));
-                        }
-                    }
-                }
-            }
-        }
-
-        // All the types that are not set will be mapped to a default type
-        // NOTE: the propagation and checks above should be enough to
-        // ensure that the unset types are never populated with any values
-        // Otherwise, since the default type is not necessarily compartible with all other types,
-        // this can lead to runtime errors
-        // TODO: Constraints are not fully covered in inference and checks at the moment
-        let result = predicate_types
-            .into_iter()
-            .map(|(predicate, types)| {
-                (
-                    predicate,
-                    types
-                        .into_iter()
-                        .map(|t| Option::<PrimitiveType>::from(t).unwrap_or_default())
-                        .collect(),
-                )
-            })
-            .collect();
-
-        Ok(result)
     }
 
     /// Check if the program contains rules with unsupported features
@@ -1058,116 +377,6 @@ impl ChaseProgram {
         Ok(())
     }
 
-    fn check_for_nonnumeric_arithmetic(&self, analyses: &[RuleAnalysis]) -> Result<(), TypeError> {
-        for (rule, analysis) in self.rules().iter().zip(analyses.iter()) {
-            for constraint in rule.all_constraints() {
-                if let Some(example_variable) = constraint.variables().next() {
-                    let variable_type = analysis
-                        .variable_types
-                        .get(example_variable)
-                        .expect("Previous analysis should have assigned a type to each variable.");
-
-                    // Note: For now, separating this two checks doesn't make much sense
-                    // because every type that allows for instance an addition, also allows comparison such as <
-
-                    if constraint.is_numeric() && !variable_type.allows_numeric_operations() {
-                        return Err(TypeError::InvalidRuleNonNumericComparison);
-                    }
-
-                    if (!constraint.left().is_primitive() || !constraint.right().is_primitive())
-                        && !variable_type.allows_numeric_operations()
-                    {
-                        return Err(TypeError::InvalidRuleNonNumericArithmetic);
-                    }
-                }
-            }
-
-            for constructor in rule.constructors() {
-                let variable_type = analysis
-                    .variable_types
-                    .get(constructor.variable())
-                    .expect("Previous analysis should have assigned a type to each variable.");
-
-                if !constructor.term().is_primitive() && !variable_type.allows_numeric_operations()
-                {
-                    return Err(TypeError::InvalidRuleNonNumericArithmetic);
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn check_aggregate_types(&self, analyses: &[RuleAnalysis]) -> Result<(), TypeError> {
-        for (rule, analysis) in self.rules().iter().zip(analyses.iter()) {
-            for aggregate in rule.aggregates() {
-                let variable_type = analysis
-                    .variable_types
-                    .get(&aggregate.variables[0])
-                    .expect("Previous analysis should have assigned a type to each aggregate output variable.");
-
-                aggregate
-                    .logical_aggregate_operation
-                    .check_input_type(&aggregate.variables[0].name(), *variable_type)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Check if there is a constant that cannot be converted into the type of
-    /// the variable/predicate position it is compared to.
-    fn check_for_incompatible_constant_types(
-        &self,
-        analyses: &[RuleAnalysis],
-        predicate_types: &HashMap<Identifier, Vec<PrimitiveType>>,
-    ) -> Result<(), TypeError> {
-        for fact in self.facts() {
-            let predicate_types = predicate_types
-                .get(&fact.predicate())
-                .expect("Previous analysis should have assigned a type vector to each predicate.");
-
-            for (term_index, constant) in fact.terms().iter().enumerate() {
-                let logical_type = predicate_types[term_index];
-                logical_type.ground_term_to_data_value_t(constant.clone())?;
-            }
-        }
-
-        for (rule, analysis) in self.rules().iter().zip(analyses.iter()) {
-            for constraint in rule.all_constraints() {
-                if let Some(example_variable) = constraint.variables().next() {
-                    let variable_type = analysis
-                        .variable_types
-                        .get(example_variable)
-                        .expect("Previous analysis should have assigned a type to each variable.");
-
-                    for term in [constraint.left(), constraint.right()] {
-                        for primitive_term in term.primitive_terms() {
-                            if let PrimitiveTerm::Constant(constant) = primitive_term {
-                                variable_type.ground_term_to_data_value_t(constant.clone())?;
-                            }
-                        }
-                    }
-                }
-            }
-
-            for constructor in rule.constructors() {
-                let variable_type = analysis
-                    .variable_types
-                    .get(constructor.variable())
-                    .expect("Previous analysis should have assigned a type to each variable.");
-
-                for term in constructor.term().primitive_terms() {
-                    if let PrimitiveTerm::Constant(constant) = term {
-                        variable_type.ground_term_to_data_value_t(constant.clone())?;
-                    }
-                }
-            }
-        }
-
-        Ok(())
-    }
-
     /// Analyze itself and return a struct containing the results.
     pub fn analyze(&self) -> Result<ProgramAnalysis, Error> {
         let BuilderResultVariants {
@@ -1178,677 +387,40 @@ impl ChaseProgram {
         let all_predicates = self.get_all_predicates();
         let derived_predicates = self.get_head_predicates();
 
-        let position_graph = self.build_position_graph();
-        let predicate_types = self.infer_predicate_types(&position_graph, &all_predicates)?;
+        let (predicate_types, rule_var_types) = infer_types(self)?;
 
         let rule_analysis: Vec<RuleAnalysis> = self
             .rules()
             .iter()
+            .zip(rule_var_types)
             .enumerate()
-            .map(|(i, r)| {
+            .map(|(i, (r, vt))| {
                 analyze_rule(
                     r,
                     all_variable_orders[i].clone(),
                     &all_column_orders,
                     i,
+                    vt,
                     &predicate_types,
                 )
             })
             .collect();
-
-        self.check_for_incompatible_constant_types(&rule_analysis, &predicate_types)?;
-        self.check_for_nonnumeric_arithmetic(&rule_analysis)?;
-        self.check_aggregate_types(&rule_analysis)?;
 
         Ok(ProgramAnalysis {
             rule_analysis,
             derived_predicates,
             all_predicates,
             predicate_types,
-            position_graph,
         })
     }
 }
 
 #[cfg(test)]
 mod test {
-    use std::collections::HashMap;
-
     use crate::{
-        io::parser::parse_program,
-        model::{
-            chase_model::{
-                ChaseFact, ChaseProgram, ChaseRule, Constructor, PrimitiveAtom, VariableAtom,
-            },
-            Constant, Constraint, DataSourceDeclaration, DsvFile, Identifier, NativeDataSource,
-            NumericLiteral, PrimitiveTerm, PrimitiveType, Term, TupleConstraint, Variable,
-        },
-        program_analysis::analysis::{get_fresh_rule_predicate, RuleAnalysisError},
+        io::parser::parse_program, model::chase_model::ChaseProgram,
+        program_analysis::analysis::RuleAnalysisError,
     };
-
-    type TestData = (
-        (ChaseRule, ChaseRule, ChaseRule),
-        (ChaseFact, ChaseFact, ChaseFact),
-        (
-            Identifier,
-            Identifier,
-            Identifier,
-            Identifier,
-            Identifier,
-            Identifier,
-            Identifier,
-        ),
-    );
-
-    fn get_test_rules_and_facts_and_predicates() -> TestData {
-        let a = Identifier("a".to_string());
-        let b = Identifier("b".to_string());
-        let c = Identifier("c".to_string());
-        let r = Identifier("r".to_string());
-        let s = Identifier("s".to_string());
-        let t = Identifier("t".to_string());
-        let q = Identifier("q".to_string());
-
-        let x = Variable::Universal(Identifier("x".to_string()));
-        let headop = Variable::Universal(Identifier("headop".to_string()));
-        let z = Variable::Existential(Identifier("z".to_string()));
-
-        let tx = PrimitiveTerm::Variable(x.clone());
-        let theadop = PrimitiveTerm::Variable(headop.clone());
-        let tz = PrimitiveTerm::Variable(z);
-
-        let v42 = Variable::Universal(Identifier("PLACEHOLDER_42".to_string()));
-        let c42 = Constant::NumericLiteral(NumericLiteral::Integer(42));
-        let t42 = PrimitiveTerm::Constant(c42.clone());
-        let tt42 = Term::Primitive(t42.clone());
-
-        let c3 = Constant::NumericLiteral(NumericLiteral::Integer(3));
-
-        let v7 = Variable::Universal(Identifier("PLACEHOLDER_7".to_string()));
-        let c7 = Constant::NumericLiteral(NumericLiteral::Integer(7));
-        let t7 = PrimitiveTerm::Constant(c7.clone());
-        let tt7 = Term::Primitive(t7.clone());
-
-        let tt55 = Term::Primitive(PrimitiveTerm::Constant(Constant::NumericLiteral(
-            NumericLiteral::Integer(55),
-        )));
-
-        // A(x) :- B(x), C(x).
-        let basic_rule = ChaseRule::new(
-            vec![PrimitiveAtom::new(a.clone(), vec![tx.clone()])],
-            vec![],
-            vec![],
-            vec![
-                VariableAtom::new(b.clone(), vec![x.clone()]),
-                VariableAtom::new(c.clone(), vec![x.clone()]),
-            ],
-            vec![],
-            vec![],
-            vec![],
-        );
-
-        // R(x, !z) :- A(x).
-        let exis_rule = ChaseRule::new(
-            vec![PrimitiveAtom::new(r.clone(), vec![tx.clone(), tz])],
-            vec![],
-            vec![],
-            vec![VariableAtom::new(a.clone(), vec![x.clone()])],
-            vec![],
-            vec![],
-            vec![],
-        );
-
-        // S(x, 55) :- T(42, x), Q(7).
-        let rule_with_constant = ChaseRule::new(
-            vec![PrimitiveAtom::new(s.clone(), vec![tx.clone(), theadop])],
-            vec![Constructor::new(headop, tt55)],
-            vec![],
-            vec![
-                VariableAtom::new(t.clone(), vec![v42.clone(), x]),
-                VariableAtom::new(q.clone(), vec![v7.clone()]),
-            ],
-            vec![
-                Constraint::Equals(Term::Primitive(PrimitiveTerm::Variable(v42)), tt42),
-                Constraint::Equals(Term::Primitive(PrimitiveTerm::Variable(v7)), tt7),
-            ],
-            vec![],
-            vec![],
-        );
-
-        let fact1 = ChaseFact::new(t.clone(), vec![c42.clone(), c3.clone()]);
-        let fact2 = ChaseFact::new(t.clone(), vec![c3, c7.clone()]);
-        let fact3 = ChaseFact::new(t.clone(), vec![c7, c42]);
-
-        (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        )
-    }
-
-    #[test]
-    fn infer_types_no_decl() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let no_decl = ChaseProgram::new(
-            None,
-            Default::default(),
-            Default::default(),
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            Default::default(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::Any]),
-            (b, vec![PrimitiveType::Any]),
-            (c, vec![PrimitiveType::Any]),
-            (r, vec![PrimitiveType::Any, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = no_decl
-            .infer_predicate_types(
-                &no_decl.build_position_graph(),
-                &no_decl.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_a_string_decl() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let a_string_decl = ChaseProgram::new(
-            None,
-            Default::default(),
-            Default::default(),
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [(a.clone(), vec![PrimitiveType::String])]
-                .into_iter()
-                .collect(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::String]),
-            (b, vec![PrimitiveType::Any]),
-            (c, vec![PrimitiveType::Any]),
-            (r, vec![PrimitiveType::String, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = a_string_decl
-            .infer_predicate_types(
-                &a_string_decl.build_position_graph(),
-                &a_string_decl.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_a_int_decl() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let a_int_decl = ChaseProgram::new(
-            None,
-            Default::default(),
-            Default::default(),
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [(a.clone(), vec![PrimitiveType::Integer])]
-                .into_iter()
-                .collect(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::Integer]),
-            (b, vec![PrimitiveType::Any]),
-            (c, vec![PrimitiveType::Any]),
-            (r, vec![PrimitiveType::Integer, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = a_int_decl
-            .infer_predicate_types(
-                &a_int_decl.build_position_graph(),
-                &a_int_decl.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_b_string_decl() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let b_string_decl = ChaseProgram::new(
-            None,
-            Default::default(),
-            Default::default(),
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [(b.clone(), vec![PrimitiveType::String])]
-                .into_iter()
-                .collect(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::String]),
-            (b, vec![PrimitiveType::String]),
-            (c, vec![PrimitiveType::Any]),
-            (r, vec![PrimitiveType::String, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = b_string_decl
-            .infer_predicate_types(
-                &b_string_decl.build_position_graph(),
-                &b_string_decl.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_b_int_decl() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let b_integer_decl = ChaseProgram::new(
-            None,
-            Default::default(),
-            Default::default(),
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [(b.clone(), vec![PrimitiveType::Integer])]
-                .into_iter()
-                .collect(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::Integer]),
-            (b, vec![PrimitiveType::Integer]),
-            (c, vec![PrimitiveType::Any]),
-            (r, vec![PrimitiveType::Integer, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = b_integer_decl
-            .infer_predicate_types(
-                &b_integer_decl.build_position_graph(),
-                &b_integer_decl.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_b_source_decl() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let b_source_decl = ChaseProgram::new(
-            None,
-            Default::default(),
-            vec![DataSourceDeclaration::new(
-                b.clone(),
-                NativeDataSource::DsvFile(DsvFile::csv_file("", TupleConstraint::from_arity(1))),
-            )],
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            Default::default(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::String]),
-            (b, vec![PrimitiveType::String]),
-            (c, vec![PrimitiveType::Any]),
-            (r, vec![PrimitiveType::String, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = b_source_decl
-            .infer_predicate_types(
-                &b_source_decl.build_position_graph(),
-                &b_source_decl.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_c_explicit_decl_overrides_source_type() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let c_explicit_decl_overrides_source_type = ChaseProgram::new(
-            None,
-            Default::default(),
-            vec![DataSourceDeclaration::new(
-                c.clone(),
-                NativeDataSource::DsvFile(DsvFile::csv_file("", TupleConstraint::from_arity(1))),
-            )],
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [(c.clone(), vec![PrimitiveType::Integer])]
-                .into_iter()
-                .collect(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::Integer]),
-            (b, vec![PrimitiveType::Any]),
-            (c, vec![PrimitiveType::Integer]),
-            (r, vec![PrimitiveType::Integer, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = c_explicit_decl_overrides_source_type
-            .infer_predicate_types(
-                &c_explicit_decl_overrides_source_type.build_position_graph(),
-                &c_explicit_decl_overrides_source_type.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_a_and_c_conflict_with_implicit_source_decl() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, _b, c, _r, _s, _t, _q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let a_and_c_conflict_with_implicit_source_decl = ChaseProgram::new(
-            None,
-            Default::default(),
-            vec![DataSourceDeclaration::new(
-                c,
-                NativeDataSource::DsvFile(DsvFile::csv_file("", TupleConstraint::from_arity(1))),
-            )],
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [(a, vec![PrimitiveType::Integer])].into_iter().collect(),
-            Default::default(),
-        );
-
-        let inferred_types_res = a_and_c_conflict_with_implicit_source_decl.infer_predicate_types(
-            &a_and_c_conflict_with_implicit_source_decl.build_position_graph(),
-            &a_and_c_conflict_with_implicit_source_decl.get_all_predicates(),
-        );
-        assert!(inferred_types_res.is_err());
-    }
-
-    #[test]
-    fn infer_types_a_and_c_conflict_with_explicit_source_decl_that_would_be_compatible_the_other_way_around(
-    ) {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, _b, c, _r, _s, _t, _q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let a_and_c_conflict_with_explicit_source_decl_that_would_be_compatible_the_other_way_around =
-            ChaseProgram::new(
-                None,
-                Default::default(),
-                vec![DataSourceDeclaration::new(
-                    c,
-                    NativeDataSource::DsvFile(DsvFile::csv_file(
-                        "",
-                        [PrimitiveType::Any].into_iter().collect(),
-                    )),
-                )],
-                vec![basic_rule, exis_rule, rule_with_constant],
-                vec![fact1, fact2, fact3],
-                [(a, vec![PrimitiveType::String])].into_iter().collect(),
-                Default::default(),
-            );
-
-        let inferred_types_res =
-            a_and_c_conflict_with_explicit_source_decl_that_would_be_compatible_the_other_way_around
-                .infer_predicate_types(
-                &a_and_c_conflict_with_explicit_source_decl_that_would_be_compatible_the_other_way_around
-                    .build_position_graph(),
-                &a_and_c_conflict_with_explicit_source_decl_that_would_be_compatible_the_other_way_around
-                    .get_all_predicates(),
-            );
-        assert!(inferred_types_res.is_err());
-    }
-
-    #[test]
-    fn infer_types_a_and_b_source_decl_resolvable_conflict() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let a_and_b_source_decl_resolvable_conflict = ChaseProgram::new(
-            None,
-            Default::default(),
-            vec![DataSourceDeclaration::new(
-                b.clone(),
-                NativeDataSource::DsvFile(DsvFile::csv_file("", TupleConstraint::from_arity(1))),
-            )],
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [(a.clone(), vec![PrimitiveType::Any])]
-                .into_iter()
-                .collect(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::Any]),
-            (b, vec![PrimitiveType::String]),
-            (c, vec![PrimitiveType::Any]),
-            (r, vec![PrimitiveType::Any, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = a_and_b_source_decl_resolvable_conflict
-            .infer_predicate_types(
-                &a_and_b_source_decl_resolvable_conflict.build_position_graph(),
-                &a_and_b_source_decl_resolvable_conflict.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_r_source_decl_resolvable_conflict_with_exis() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let r_source_decl_resolvable_conflict_with_exis = ChaseProgram::new(
-            None,
-            Default::default(),
-            vec![DataSourceDeclaration::new(
-                r.clone(),
-                NativeDataSource::DsvFile(DsvFile::csv_file("", TupleConstraint::from_arity(2))),
-            )],
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            Default::default(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::Any]),
-            (b, vec![PrimitiveType::Any]),
-            (c, vec![PrimitiveType::Any]),
-            (r, vec![PrimitiveType::String, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = r_source_decl_resolvable_conflict_with_exis
-            .infer_predicate_types(
-                &r_source_decl_resolvable_conflict_with_exis.build_position_graph(),
-                &r_source_decl_resolvable_conflict_with_exis.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
-
-    #[test]
-    fn infer_types_b_and_c_conflict_decl() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (_a, b, c, _r, _s, _t, _q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let b_and_c_conflict_decl = ChaseProgram::new(
-            None,
-            Default::default(),
-            Default::default(),
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [
-                (b, vec![PrimitiveType::Integer]),
-                (c, vec![PrimitiveType::String]),
-            ]
-            .into_iter()
-            .collect(),
-            Default::default(),
-        );
-
-        let inferred_types_res = b_and_c_conflict_decl.infer_predicate_types(
-            &b_and_c_conflict_decl.build_position_graph(),
-            &b_and_c_conflict_decl.get_all_predicates(),
-        );
-        assert!(inferred_types_res.is_err());
-    }
-
-    #[test]
-    fn infer_types_b_anc_c_conflict_decl_resolvable() {
-        let (
-            (basic_rule, exis_rule, rule_with_constant),
-            (fact1, fact2, fact3),
-            (a, b, c, r, s, t, q),
-        ) = get_test_rules_and_facts_and_predicates();
-
-        let b_and_c_conflict_decl_resolvable = ChaseProgram::new(
-            None,
-            Default::default(),
-            Default::default(),
-            vec![basic_rule, exis_rule, rule_with_constant],
-            vec![fact1, fact2, fact3],
-            [
-                (b.clone(), vec![PrimitiveType::Any]),
-                (c.clone(), vec![PrimitiveType::String]),
-            ]
-            .into_iter()
-            .collect(),
-            Default::default(),
-        );
-
-        let expected_types: HashMap<Identifier, Vec<PrimitiveType>> = [
-            (a, vec![PrimitiveType::Any]),
-            (b, vec![PrimitiveType::Any]),
-            (c, vec![PrimitiveType::String]),
-            (r, vec![PrimitiveType::Any, PrimitiveType::Any]),
-            (s, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (t, vec![PrimitiveType::Integer, PrimitiveType::Integer]),
-            (q, vec![PrimitiveType::Integer]),
-            (get_fresh_rule_predicate(1), vec![PrimitiveType::Any]),
-        ]
-        .into_iter()
-        .collect();
-
-        let inferred_types = b_and_c_conflict_decl_resolvable
-            .infer_predicate_types(
-                &b_and_c_conflict_decl_resolvable.build_position_graph(),
-                &b_and_c_conflict_decl_resolvable.get_all_predicates(),
-            )
-            .unwrap();
-        assert_eq!(inferred_types, expected_types);
-    }
 
     #[test]
     #[cfg_attr(miri, ignore)]
